More settings work
Settings can be written and read, prep work for the Error handling
mechanism.
